--- Assets/Scripts/Field/Data/StageData.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 29225e3ec3c85d443a47cc7660c38cf2
