--- Assets/Scripts/Field/FieldTest/Data/SO_Field.asset
@@ -1,22 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 29225e3ec3c85d443a47cc7660c38cf2, type: 3}
-  m_Name: SO_Field
-  m_EditorClassIdentifier: Assembly-CSharp::StageData
-  _width: 20
-  _height: 20
-  FieldGroundPrefab: {fileID: 7577194185336080069, guid: 029d45697f8b03c4eace89d1107069f8, type: 3}
-  spawns:
-  - SpawnerPrefab: {fileID: 5284306673753496953, guid: 6bd9455dc82e5e744aceaecaa564c6db, type: 3}
-    MockPrefab: {fileID: 4933468142118675558, guid: 62d601492dbcb294783b4307fd092bbc, type: 3}
-    position: {x: 0, y: 2, z: 0}
-    SpawnInterval: 0
