--- Assets/Scripts/Field/FieldTest/FieldTestSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ac32dab66bc881442900e927592b3e17
