+++ Assets/Scripts/Gameplay/Stage/Field/Data/StageData.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 29225e3ec3c85d443a47cc7660c38cf2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - FieldGroundPrefab: {instanceID: 0}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
