+++ Assets/Scripts/Gameplay/Stage/Field/StageController.cs
@@ -1,0 +1,23 @@
+using UnityEngine;
+
+namespace Game.Gameplay.Stage
+{
+
+    public class StageController
+    {
+        [Header("生成するオブジェクト")]
+
+        [SerializeField]
+        StageData _stageData;
+
+        public StageData StageData => _stageData;
+
+        public void Initialize(StageData stageData)
+        {
+            _stageData = stageData;
+
+        
+        }
+    }
+
+}
