+++ Assets/_Project.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 4b2e90e0b3a7c2b4da7a64457c69134e
-NativeFormatImporter:
+guid: 09c48ee312193414ea5f043cd8433bed
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
