+++ Assets/_Project/Editor/Game.Editor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 328a82ac225f7f54cbf7d5054a518a8f
-folderAsset: yes
-DefaultImporter:
+guid: 5afdd5e23f82fa749b04bd5077175ade
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
