+++ Assets/_Project/Runtime/Core.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 62d601492dbcb294783b4307fd092bbc
-PrefabImporter:
+guid: 613e8f71b00007649b32a207aa5f6f20
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
