+++ Assets/_Project/Runtime/Core/ICoreContracts.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 93f44b60abd778d41ab1a8aff0ddac26
