+++ Assets/_Project/Runtime/Events.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 029d45697f8b03c4eace89d1107069f8
-PrefabImporter:
+guid: 2d5a15236256e0d41930f47e0570db3b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
