+++ Assets/_Project/Runtime/Events/EventBus.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ba9b522676d5039448a5fabe08b4f049
