+++ Assets/_Project/Runtime/Game.Runtime.asmdef.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: b9c53ecdde96a1146ad64ed272b2c5ff
-NativeFormatImporter:
+guid: a7e50a02167266d4e861852bf8ea8696
+AssemblyDefinitionImporter:
   externalObjects: {}
-  mainObjectFileID: 13400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
