+++ Assets/_Project/Tests/Game.Tests.asmdef.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6bd9455dc82e5e744aceaecaa564c6db
-PrefabImporter:
+guid: 624c28bd498c1ce44af2115c2c82833e
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
